Get rid of low mode button mode and make thresholds actually work (#606)
* Get rid of low mode button and increase button thresholds

* Hanson is incredibly dumb

* Remove useless cout

+++ muan/teleop/button.h
@@ -90,7 +90,7 @@ class AxisRange : public Button {
   int minimum_;
   int maximum_;
   int yaxis_;
-  int threshold_;
+  double threshold_;
 };
 
 }  // namespace teleop
